refactor rrv entities

+++ lib/aca_entities/fdsh/rrv/h79/batch_metadata.rb
@@ -8,9 +8,9 @@ module AcaEntities
         class BatchMetadata < Dry::Struct
           attribute :BatchID,  AcaEntities::Types::String.optional.meta(omittable: false)
           attribute :BatchPartnerID,  AcaEntities::Types::String.optional.meta(omittable: false)
-          attribute :BatchAttachmentTotalQuantity,  Types::Float.optional.meta(omittable: false)
+          attribute :BatchAttachmentTotalQuantity,  Types::Integer.optional.meta(omittable: false)
           attribute :BatchCategoryCode,  AcaEntities::Types::String.optional.meta(omittable: false)
-          attribute :BatchTransmissionQuantity,  Types::Float.optional.meta(omittable: false)
+          attribute :BatchTransmissionQuantity,  Types::Integer.optional.meta(omittable: false)
         end
       end
     end
